+++ documentation/changelog.rst
@@ -8,6 +8,7 @@ v0.33.0 | May XX, 2026
 
 New features
 -------------
+* Improve UX after deleting a child asset through the UI [see `PR #2119 <https://www.github.com/FlexMeasures/flexmeasures/pull/2119>`_]
 
 Infrastructure / Support
 ----------------------